Update tuple tests
There's something wierd going on with the types here though
